pyproject.toml
[project]
name = "find-image-in-pdf"
version = "0.1.0"
description = "Find which page of a PDF contains a given image"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pymupdf>=1.23.0",
    "pillow>=10.0.0",
    "imagehash>=4.3.0",
    "gradio>=4.0.0",
]

[project.scripts]
find-image = "find_image_in_pdf:main"
find-image-gui = "find_image_in_pdf_gui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
